Reduce tech debt
