+++ app/globals.css
@@ -1,5 +1,4 @@
 @import "tailwindcss";
-@import "tw-animate-css";
 
 /* Dotted background utility */
 .bg-dots {
