+++ package.json
@@ -56,7 +56,6 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.59.0",
-    "recharts": "^3.1.2",
     "resend": "^4.6.0",
     "server-only": "^0.0.1",
     "sonner": "^2.0.7",
@@ -66,21 +65,13 @@
   "devDependencies": {
     "@eslint/eslintrc": "^3",
     "@playwright/test": "^1.54.1",
-    "@tailwindcss/postcss": "^4",
     "@testing-library/jest-dom": "^6.6.3",
-    "@types/jest": "^30.0.0",
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
-    "eslint": "^9",
-    "eslint-config-next": "15.3.4",
     "jest": "^30.0.5",
-    "jest-environment-jsdom": "^30.0.5",
     "prettier": "^3.3.3",
     "prisma": "^6.10.1",
-    "tailwindcss": "^4",
-    "ts-jest": "^29.4.0",
-    "tw-animate-css": "^1.3.4",
     "typescript": "^5"
   }
 }
